feat(terminal): handle ReaderListener's onDisconnect
